pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paprwall"
version = "1.0.0"
description = "Auto-rotating wallpaper app for Linux with multi-source support and local images"
authors = [{name = "riturajprofile", email = "riturajprofile.me@gmail.com"}]
license = {text = "CC BY-NC 4.0"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["wallpaper", "linux", "desktop", "pixabay", "unsplash", "pexels", "paprwall"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: Other/Proprietary License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Desktop Environment",
]

dependencies = [
    "requests>=2.28.0",
    "Pillow>=10.0.0",
    "APScheduler>=3.10.0",
    "PyGObject>=3.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/riturajprofile/paprwall"
Documentation = "https://github.com/riturajprofile/paprwall/wiki"
Repository = "https://github.com/riturajprofile/paprwall"
"Bug Tracker" = "https://github.com/riturajprofile/paprwall/issues"

[project.scripts]
paprwall = "riturajprofile_wallpaper.cli:main"

[project.gui-scripts]
paprwall-gui = "riturajprofile_wallpaper.gui.main_window:main"
